+++ src/dtext.js
@@ -12,9 +12,15 @@ export default class DText {
       </div>
     `);
 
-    $expandable.find('.expandable-button').click(e => {
-      $(e.target).closest('.expandable').find('.expandable-content').fadeToggle('fast');
-      $(e.target).val((_, val) => val === 'Show' ? 'Hide' : 'Show');
+    // If our script runs before Danbooru's scripts do, Danbooru will find our
+    // expandable and add it's own click handler on top of ours. So delay
+    // adding our handler to make sure we overwrite Danbooru's own handler.
+    $(function () {
+      $expandable.find('.expandable-button').off("click").click(e => {
+        $(e.target).closest('.expandable').find('.expandable-content').fadeToggle('fast');
+        $(e.target).val((_, val) => val === 'Show' ? 'Hide' : 'Show');
+        e.preventDefault();
+      });
     });
 
     return $expandable;